requirements.txt
streamlit
pandas
google-generativeai
python-dotenv
psycopg2-binary
SQLAlchemy
matplotlib
tabulate